requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
cmake==3.31.0.1
idna==3.10
numpy==2.1.3
opencv-face-recognition==1.1.1
opencv-python==4.10.0.84
pillow==11.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
shiboken6==6.8.0.2
six==1.16.0
urllib3==2.2.3
